[WIP] The Soda-specific code has been moved into a dedicated Go source file
